Update W000820
For some reason there are two actions that handle the ZIP submission click - 'click_on' and a javascript click. This is causing errors when the first click works.

Not sure why there are two actions handling this, I believe one is all you need. However, if two are need for some other reason , I have added some logic to check if the button has already been clicked to prevent it timing out.